fix some warnings
This call site is reachable on: 'iOS' 12.2 and later, 'maccatalyst' 12.2 and later. 'AVCaptureDeviceType.BuiltInUltraWideCamera' is only supported on: 'ios' 13.0 and later, 'maccatalyst' 13.0 and later, 'tvos' 17.0 and later.
This call site is reachable on: 'iOS' 12.2 and later, 'maccatalyst' 12.2 and later. 'AVCaptureDeviceType.BuiltInDualWideCamera' is only supported on: 'ios' 13.0 and later, 'maccatalyst' 13.0 and later, 'tvos' 17.0 and later.
This call site is reachable on: 'iOS' 12.2 and later, 'maccatalyst' 12.2 and later. 'AVCaptureDeviceType.BuiltInTripleCamera' is only supported on: 'ios' 13.0 and later, 'maccatalyst' 13.0 and later, 'tvos' 17.0 and later.

* fixup for commit 199e743265bb61f331025de31e95e4af93afb16a